requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic[email]>=2.10.0
python-dotenv>=1.0.0
pymysql>=1.1.0
PyJWT>=2.8.0
bcrypt>=4.1.2
python-multipart>=0.0.6
httpx>=0.26.0
